fix: 10818
git-svn-id: http://halodev.ontoprise.com/svn/HaloSMWExtension/trunk@4536 f80b7aa0-dfcc-4e65-a0d9-74108174cbc8

+++ extensions/SemanticGardening/includes/storage/SGA_GardeningLogSQL.php
@@ -251,8 +251,8 @@ class SGAGardeningLogSQL extends SGAGardeningLog {
 			while($row = $dbr->fetchObject($res))
 			{
 				// get name of log page and remove the article
-				$log = explode("/", $row->log);
-				$logTitle = Title::newFromDBkey($log[count($log)-1]);
+				$logTitle = Title::newFromDBkey($row->log);
+				if (is_null($logTitle) || !$logTitle->exists()) continue;
 				$logArticle = new Article($logTitle);
 				if ($logArticle->exists()) {
 					$logArticle->doDeleteArticle("automatic deletion");
